switch FIAT -> FInAT for point evaluation
