+++ firedrake/function.py
@@ -606,7 +606,7 @@ def make_c_evaluate(function, c_name="evaluate", ldargs=None, tolerance=None):
 
     mesh = function.ufl_domain()
     src = pq_utils.src_locate_cell(mesh, tolerance=tolerance)
-    src += compile_element(function)
+    src += compile_element(function, mesh.coordinates)
 
     args = []
